temp

+++ src/Exceptions/NotEnoughCoins.java
@@ -3,7 +3,7 @@ package Exceptions;
 import Players.Player;
 
 public class NotEnoughCoins extends Exception {
-    public NotEnoughCoins(Player player, int unitCost) {
+    public NotEnoughCoins(Player player, Double unitCost) {
         super("You do not have enough coins! You have " + player.getCoins() + " coins, but you need " + unitCost + " coins!");
     }
 }
